+++ MyFunctions/internal/configurations/configuration.ps1
@@ -18,6 +18,12 @@ Set-PSFConfig -Module 'MyFunctions' -Name 'DHCP.FQDN' -Value '' -Initialize -Val
 Set-PSFConfig -Module 'MyFunctions' -Name 'IMC.FQDN' -Value '' -Initialize -Validation 'String' -Handler { } -Description "IMC Server der per Get-ImcSwitchPort abgefragt wird." -ModuleExport
 Set-PSFConfig -Module 'MyFunctions' -Name 'SMTP.From' -Value '' -Initialize -Validation 'String' -Handler { } -Description "Absender Adresse die bei der Nutzung von Send-MailMessage in verschiedenen CmdLets genutzt wird." -ModuleExport
 Set-PSFConfig -Module 'MyFunctions' -Name 'SMTP.Server' -Value '' -Initialize -Validation 'String' -Handler { } -Description "SMTP Server der bei der Nutzung von Send-MailMessage von verschiedenen CmdLets genutzt wird." -ModuleExport
+Set-PSFConfig -Module 'MyFunctions' -Name 'PhoneList.ExportPath' -Value '' -Initialize -Validation 'String' -Handler { } -Description "Pfad unter dem die Telefonlisten Exporte von iDoit abgelegt werden" -ModuleExport
+Set-PSFConfig -Module 'MyFunctions' -Name 'PhoneList.RemotePath' -Value '' -Initialize -Validation 'String' -Handler { } -Description "Remotepfad unter dem die Telefonlisten Exporte hochgeladen werden" -ModuleExport
+Set-PSFConfig -Module 'MyFunctions' -Name 'PhoneList.RemoteServer' -Value '' -Initialize -Validation 'String' -Handler { } -Description "Intranetserver auf dem die Telefonliste veröffentlicht wird" -ModuleExport
+Set-PSFConfig -Module 'MyFunctions' -Name 'PhoneList.KeyFile' -Value '' -Initialize -Validation 'String' -Handler { } -Description "Pfad zur OpenSSH Schlüsseldatei die für die SSH Verbindung genutzt wird." -ModuleExport
+Set-PSFConfig -Module 'MyFunctions' -Name 'PhoneList.SSHUser' -Value '' -Initialize -Validation 'String' -Handler { } -Description "User der für die SSH Verbindung genutzt wird." -ModuleExport
+
 
 If ($ConfigurationFile)
 {